refactor: 프로필 페이지 컴포넌트 구조 변경
