+++ src/constants/window-view/profile-window-view.tsx
@@ -6,7 +6,7 @@ export const PROFILE_PAGE_ID = 'PROFILE'
 export const PROFILE_PAGE_VIEW: Omit<WindowViewProps, 'id'> = {
   title: '프로필',
   closable: true,
-  defaultSize: { width: 700, height: 500 },
+  defaultSize: { width: 550, height: 500 },
   minWidth: 350,
   minHeight: 500,
   maxWidth: 1065,
