+++ src/pages/profile/components/index.ts
@@ -1,5 +1,5 @@
-export * from './streak-date-selector'
-export * from './streak-type-selector'
-export * from './daily-streak'
-export * from './weekly-streak'
-export * from './streak'
+export * from './streak/streak-date-selector'
+export * from './streak/streak-type-selector'
+export * from './streak/daily-streak'
+export * from './streak/weekly-streak'
+export * from './streak/streak'
