+++ src/pages/profile/components/streak/streak-type-selector.tsx
