+++ src/pages/profile/components/streak/weekly-streak.tsx
@@ -1,11 +1,11 @@
 import { useStreak } from '@/api/hooks'
 import { CalendarHeatmap } from '@/components/heatmap'
-import { StreakDateSelector } from '@/pages/profile/components/streak-date-selector'
+import { StreakDateSelector } from '@/pages/profile/components/streak/streak-date-selector'
 import { alpha, Divider, Stack } from '@mui/material'
 import dayjs from 'dayjs'
 import { useMemo, useState } from 'react'
 
-export function DailyStreak() {
+export function WeeklyStreak() {
   const [endDate, setEndDate] = useState(new Date())
 
   const startDate = useMemo(() => {
@@ -23,11 +23,11 @@ export function DailyStreak() {
       height={150}
     >
       <CalendarHeatmap
-        data={streak.daily}
+        data={streak.weekly}
         endDate={endDate}
         sx={{ flex: 1, overflow: 'hidden', p: 1, alignSelf: 'center' }}
         color="#00a63e"
-        type="DAILY"
+        type="WEEKLY"
       />
       <Divider orientation="vertical" />
       <StreakDateSelector value={endDate} onChange={setEndDate} sx={{ width: 80, p: 0.5 }} />
